+++ src/org/linkeddatafragments/datasource/AbstractRequestProcessorForTriplePatterns.java
@@ -1,5 +1,11 @@
 package org.linkeddatafragments.datasource;
 
+import org.linkeddatafragments.fragments.LinkedDataFragment;
+import org.linkeddatafragments.fragments.LinkedDataFragmentRequest;
+import org.linkeddatafragments.fragments.TPF.TriplePatternFragment;
+import org.linkeddatafragments.fragments.TPF.TriplePatternFragmentImpl;
+import org.linkeddatafragments.fragments.TPF.TriplePatternFragmentRequest;
+
 import com.hp.hpl.jena.rdf.model.Model;
 import com.hp.hpl.jena.rdf.model.Property;
 import com.hp.hpl.jena.rdf.model.RDFNode;
