+++ src/org/linkeddatafragments/datasource/IFragmentRequestProcessor.java
@@ -2,6 +2,8 @@ package org.linkeddatafragments.datasource;
 
 import java.io.Closeable;
 
+import org.linkeddatafragments.fragments.LinkedDataFragment;
+
 /**
  * Processes a single request sent to a Linked Data Fragments interface.
  *
