+++ src/org/linkeddatafragments/fragments/LinkedDataFragmentBase.java
@@ -1,4 +1,4 @@
-package org.linkeddatafragments.datasource;
+package org.linkeddatafragments.fragments;
 
 import java.net.URISyntaxException;
