+++ src/org/linkeddatafragments/servlet/TriplePatternFragmentServlet.java
@@ -25,12 +25,12 @@ import org.linkeddatafragments.datasource.IDataSource;
 import org.linkeddatafragments.datasource.IFragmentRequestProcessor;
 import org.linkeddatafragments.datasource.IndexDataSource;
 import org.linkeddatafragments.datasource.JenaTDBDataSourceType;
-import org.linkeddatafragments.datasource.LinkedDataFragment;
-import org.linkeddatafragments.datasource.LinkedDataFragmentRequest;
-import org.linkeddatafragments.datasource.LinkedDataFragmentRequestBase;
-import org.linkeddatafragments.datasource.TriplePatternFragmentRequestImpl;
 import org.linkeddatafragments.exceptions.DataSourceException;
 import org.linkeddatafragments.exceptions.DataSourceNotFoundException;
+import org.linkeddatafragments.fragments.LinkedDataFragment;
+import org.linkeddatafragments.fragments.LinkedDataFragmentRequest;
+import org.linkeddatafragments.fragments.LinkedDataFragmentRequestBase;
+import org.linkeddatafragments.fragments.TPF.TriplePatternFragmentRequestImpl;
 import org.linkeddatafragments.util.MIMEParse;
 
 /**
